requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
pdfplumber==0.9.0
pytesseract==0.3.10
pdf2image==1.16.3
transformers==4.30.2
torch==2.0.1
Pillow==10.0.0
opencv-python==4.8.0.74
gunicorn==21.2.0
